Completed ".obj to .ppm" functionality.
Finished the functionality that allows me to visualize the array of information that
I will be using to generate a wav file as a .ppm file. For my purposes, the
.ppm file is a debugging file.

Up to date for both MacOS and Windows.Also, added new "temp" folder to
which a "log.txt" file is written so that I can look over the vertice/uv
data and corresponding vertex/uv map if need be.